requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
colorlog==6.8.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
hikari==2.0.0.dev122
hikari-lightbulb==2.3.5
httpcore==1.0.4
httpx==0.27.0
idna==3.6
multidict==6.0.5
openai==1.12.0
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
yarl==1.9.4